functions/dlp_scan/requirements.txt
google-cloud-dlp==3.12.3
google-cloud-storage==2.13.0
